A: Fix unnecessary fragment recreation on orientation change
The MultiStack was recreated every orientation change, now it is
scoped to the application.
