+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -39,8 +39,8 @@ object Versions {
     const val MATERIAL = "1.2.0-beta01"
     const val CONSTRAINT_LAYOUT = "2.0.0-beta4"
     const val LIFECYCLE = "2.2.0"
-    const val SIMPLE_STACK = "2.4.0"
-    const val SIMPLE_STACK_EXTENSIONS = "2.0.1"
+    const val SIMPLE_STACK = "2.5.0"
+    const val SIMPLE_STACK_EXTENSIONS = "2.1.0"
     const val RETROFIT = "2.9.0"
     const val RETROFIT_KOTLINX_SERIALIZATION = "0.8.0"
     const val OKHTTP = "4.8.0"
